dagger uses stubbed remote data sources for android instrumented tests. orchestrator added to clear app storage before each test.

+++ app/src/main/java/com/example/piggybank/dao/ExpenseDao.kt
@@ -66,9 +66,10 @@ interface ExpensesDao {
 
     @Transaction
     suspend fun overrideExpensesAndSum(expenses: List<ExpenseEntity>): Double {
-        deleteAll()
+        if (expenses.isNotEmpty()) {
+            deleteAll()
+        }
         insert(expenses)
         return sumExpenses()
     }
-
 }
